Added registered object profiler
Added an object profiler that tracks the number of instances of
registered types.  It outputs to dlog every N seconds, where N comes
from the value of the PROFILE_DALI_OBJECTS environment variable. If
the variable is unset, then no profiling will occur.

Note, the profiler relies on Dali::Stage being instantiated in order
to access the ObjectRegistry. This means it doesn't count the objects
that Stage creates (the first camera and the root layer).

Change-Id: I2af264cc4889e508dfab06a7f1dedf63844c0cc2
Signed-off-by: David Steele <david.steele@partner.samsung.com>