Added ASTC Native file format loader
ASTC can be wrapped in a KTX file, but also has its own native file format.
This patch provides support for the native ASTC file format (.astc).

Textures can be generated using tools (Eg. Mali texture compression tool).

Note: ASTC is "OpenGL friendly" in that the textures are pre-flipped vertically.
In DALi there is currently not a differenciation between different texture compression formats in that the data (once the header is extracted) is given to the hardware as-is.
For this reason textures must currently be pre-flipped vertically (before compression) when using the native .astc file format with DALI.

An example .astc file is included as part of the unit tests.

Change-Id: I574961bc184d68d3ec722be3c682ca2a2b7ed0ce