[3.0]Fix for libjpeg API changes
TurboJPEG API functions will now return an error code if a warning is triggered in the underlying libjpeg API.

For instance, if a JPEG file is corrupt, the TurboJPEG decompression functions will attempt to decompress as much of the image as possible, but those functions will now return -1 to indicate that the decompression was not entirely successful.

For this reason, we need to distinguish return of tjGetErrorStr() is warning or error.

If the return string has 'Corrupt JPEG data' prefix, it means warning.
Then, do not return false.

Change-Id: Ib542f2365c0e35c14187e3daa7e32008296c26bb