EventHandlers use IsDeviceButton for filtering Keyevents to IMF
A comparison was being done with the Ecore KeyEvent and the lookup table in in Dali KeyLookup,
as there could be one to many mappings the first match was returned. Meant a following match that was needed got ignored.
Use IsDeviceButton now which returns true or false after checking whole KeyLookup table.

Change-Id: I0d888a641e700569d6a66106df52ccef01f86a1d