Adding scale and crop feature to alpha mask
When alpha masking an image, there is a requirement to first
scale the content image by some factor, then crop it to match the
mask size, then apply the mask to it.

Uses Lanczos resampling for reducing an image, or Mitchell resampling for
upscaling an image ( Mitchell has fewer artifacts on my test images ).

Change-Id: I0334baa855a3e1547947f0a81b42f5e6271232ce
Signed-off-by: David Steele <david.steele@samsung.com>