Add locale property setting
In the font-config of the TV, the priority is generally set in the order
of Korean -> Chinese font, but in the case of zh_CN, the priority is set
to Chinese -> Korean.

In the case of EFL and Clutter, font preference policy in Font-Config is
read according to setlocale information when setting font-family.
This part is missing in DALi.

Change-Id: I8d346eed5f7935f57c6c3267ef9e7b6c164fe516