Multiply the Alpha of mask to the premultiplied image
After move MultiplyColorByAlpha() from main thread to resource thread, MultiplyColorByAlpha() is called first than ApplyMask().
In this case, the alpha value of mask is not multiplied properly to other channels in the image.
Therefore, if image is premultiplied, the other channels of the image need to multiply by alpha of mask.

Change-Id: I60c056460537db604566f080302be09a6f88d047