Fix a surface resize issue
The thread controller will wait for the response that the surface is resized in the updated thread.
But the message may be processed right after ProcessCoreEvents().
If it is processed before calling to mThreadController->ResizeSurface(), the thread controll will wait for the response forever.

So I've changed to call to mThreadController->ResizeSurface() before ProcessCoreEvents().

Change-Id: Id7d4e2a801c1d2fc8d4992676ae2b673eeb8baf0