Ensure curl redirection and proxy environment are honoured
Web redirection is ignored with the current curl settings.
Also, although curl is supposed to utilize http_proxy variable
in practice, it is ignoring it. Added code to check if it is set,
if so, use it.

Change-Id: I7a1d04bea3759aedeaeccd8c348be04d11b51a92
(cherry picked from commit 75d05ea14b1bead87a189f9bf388f8d45ecbcb87)