windows: Move getopt.h include file to the bottom
getopt-win32 defines the symbol _END_EXTERN_C and undefines at the end
of the file. It turns out msvc standard library defines this same
symbol.

If getopt.h is included before C++ standard includes, it will undef
_END_EXTERN_C and causes compilation errors under msvc.

We move the getopt.h to the end of the list of include files so it can't
mess up with internal msvcrt symbols.

Change-Id: I404ca48d096703d0be6b5f466402fb2ca0223033