platform-implement-win.cpp: Fix invalid pointer cast in 64 bits systems
The ID returned by the function SetTimer is actually a pointer to a data
structure storing the timer information. Originally the return type of
the function was int, which is 32 bits, but in Windows 64 bits it is not
enough to store a pointer value.

We convert the timer ID type to the proper type intptr_t, which is an
integral type guaranteed to be big enough to hold a pointer value.

Change-Id: I207b8e2538d73d583ed6df1192af0fbbdb6bed96