Support for the macOS platform
This commit provides basic support for the macOS platform.

The macOS specific bits are in part implemented in Objectve-C++ under
the Cocoa Framework.

One note is that we installed libANGLE through vcpkg, and it is using
CGL as the default backend. We should fix this in the future in favor of
the Metal Framework.

We added a new build profile called macos, and the cmake options
"-DENABLE_PROFILE=MACOS -DPROFILE_LCASE=macos" must be passed in the
command line to build it successfully.

cairo and fontconfig packages must be installed through brew, the
version provided by vcpkg doesn't work on macOS (actually, only
fontconfig is necessary, but if you install cairo through vcpkg, it will
pull fontconfig as well).

Change-Id: Ibcf4f854721a48e3ae4e5bd710182d0dae893e05