Initial refactoring of graphics interface
Wrapped GL abstractions in EglGraphicsController.

Cleaned up initialization of EglGraphics, moved most of it to render
thread.

(Kept initialization of GlImplementation in event thread - it has no
dependence on EGL)

TestGraphicsApplication creates EglGraphicsController, so that we can
now write tests for the new graphics API that run the actual code, and
can be tested using the existing gl abstraction.

Change-Id: I4237f52ff4912e8eec0dfc161b33e44d2f5e620f