Fix the race condition issue of RenderOnce() in GlWindow
Fix the race condition issue of RenderOnce() in GlWindow.
If RenderOnce() function is called in event thread continuously,
the race condition issue will be occured.
Because the renderOnce flag is written in event thread
and the flag is read and written in render thread without lock.
This patch is to fix the issue.

Change-Id: I0a45c0e538c19b9c552d75fd1cf1e2ca4cbd042a