atspi: suppress reading of screen-reader
Because there is legacy code for screen-reader, an application could want
to keep previous behavior of the screen-reader even though the ATSPI is
enabled by AT client such as Aurum test framework.

If an application wants to make screen-reader AT client do not read,
then set suppress to true. The default suppress is false.

Change-Id: I2a639b50d7920f59366def7b6467fe38bc754fbe