[Tizen] Modify window configure notification event callback
Modify configure notification event callback for window is resized or moved.
This evnet is sent from the display server.
If the event has 0'size, the window is resized by client application.
In this case, dali should ignore that.
If the event has real size size, the window is resized by display server.
In this case, dali should resize own's window buffer and re-render his contents.

Change-Id: Ide3d7c0fb2ad133b444a9bc66f8281c6c2906c45