Adding GLES::SyncObject implementation
The GLES::SyncObject wraps the existing EglSyncObject
(potential for refactoring here).

The old Integration::GlSyncAbstraction has been removed.

When EndRenderPass is executed, if there is a sync object,
then an EGL sync object is created through the extension API.
(This could be changed to GlCreateSyncObject with GLES 3.2)

This object is polled through the graphics API by render manager.

Change-Id: I879e344ff04f0e68c3c492e5d741a1550bd8992d
Signed-off-by: David Steele <david.steele@samsung.com>