Makes the LTR/RTL alignment of text follow the system language by default.
The MATCH_SYSTEM_LANGUAGE_DIRECTION property defaults to true.

When the layout direction of TextLabel, TextField, or TextEditor is changed,
LTR/RTL alignment of text follows the layout direction of each control.

If you set MATCH_SYSTEM_LANGUAGE_DIRECTION to false
As before, LTR/RTL is determined according to the language of the text.

Change-Id: If92e44feb942cf3024bbe56842d75b2b45a73448