[AT-SPI] Remove Role check in AcceptObject
Removed AcceptObjectCheckRole() method and its call in AcceptObject().

Checking the AT-SPI Role of the Component to decide whether it should be
highlighted or not is not needed, as by default all Components are
non-highlightable and highlightability is decided individually
in dali-toolkit by setting the ACCESSIBILITY_HIGHLIGHTABLE property.

From now on if we want to turn on highlightability for certain
Components, we can do this in dali-toolkit on a class level
or in runtime per individual instance of a Component.

Change-Id: I4777594550765b0b8335e47cd322d7d96590fe55