[ATSPI] Read property in Async way
An application could launch and read a property before at-spi-bus-launcher
is running. The at-spi-bus-launcher is a process writing the property.
If the application reads the property in syncronous way, then it is waiting
until the at-spi-bus-launcher can respond.

For the performance reason we will not read properties in syncronous way.

Change-Id: I0c7f33b25f99d37d57c67bca48e2e617ab5b1b1a