[AT-SPI] Add Bridge::{Enabled,Disabled}Signal()
These signals are emitted whenever the AT-SPI bridge is turned on or
off. Note that `EnabledSignal` is only emitted for the first `ForceUp`
call, i.e. the one that returns `ForceUpResult::JUST_STARTED` (and
a similar story for `DisabledSignal`).

These signals are necessarily static members of `Bridge`, to guarantee
that they are delivered to consumers even in those use cases in which
the object returned by `GetCurrentBridge()` changes during the lifetime
of the application.

Change-Id: Id266b7ed064b8b2690e093804c3f7d23ac4347ce