[AT-SPI] Clean up AT-SPI interfaces
This patch adds a `const` qualifier to most getter methods across all
AT-SPI interfaces, with the exception of those that return an
`Accessible*`, since these objects are by nature mutable (i.e. exposed
on DBus for external manipulation, and DBus does not have const
objects), and there is always a possibility to return to `this` with the
`const` cast away, e.g. GetParent()->GetChildAtIndex(i).

There are also some other minor changes like a `noexcept` qualifier for
`~Accessible`, the `EditableText` interface now implying `Text`, and a
data-driven `GetRoleName` in `Accessible`.

Change-Id: I2ab0bb451f188fe0bd6dbd47c4aa0bad56624b42