[ATSPI] ignore ScreenReaderEnabled property on suppress mode
The 'suppress' mode means that we do not want screen-reader working
for our ATSPI related event. In this case, it would be better NOT to
enable ATSPI bridge.

So far we have used 'suppress-screen-reader' attribute. But sometimes
GetAttributes does not work with following error message.

"get attribute error: Did not receive a reply. Possible causes include:
 the remote application did not send a reply, the message bus security
 policy blocked the reply, the reply timeout expired, or the network
 connection was broken."

We could handle this case, but the best is NOT to send unnecessary
ATSPI events.

The screen-reader should turning on ScreenReaderEnabled property only
The ATSPI bridge will be enabled, when IsEnabled property is set by
another AT client such as Aurum.

Change-Id: I529ae8cc29594915b20c23279371a6488d11ea2d