Delayed SceneHolder RenderTarget initialization
Initialization of Surface/Context pair happened in wrong thread: for
SceneHolder, the RenderTarget was being created in the Event thread,
which has effect of modifying a render-thread owned vector / queue
in the event thread, which causes crash.

Instead, moved the RenderTarget creation to
SceneGraph::Scene::Initialize(), which is executed in the render
thread.

Change-Id: Ifac94b08b4cb099267abc0353a995539ed5493c2
Signed-off-by: David Steele <david.steele@samsung.com>