[AT-SPI] Rework children handling in ActorAccessible
This patch reorganizes the children-related methods in ActorAccessible.
A few methods are also marked 'final' to prevent overriding in derived
classes. A new overridable method, DoGetChildren() is provided instead,
to offer a simple and bulletproof mechanism for derived classes to
customize the children to be reported in the AT-SPI tree (possibly
including non-ActorAccessible objects).

Change-Id: I0c34c3493091c108e94aa140f883f012aea347f6