Add SyncKeyEvent for Widget
In the past, the WidgetApplication used the keyEvent using the screen connector's API.
However, this API is not synchronized.
So WidgetView cannot get feedback from WidgetApplication.
To solve this problem, AppFramework provides a new API to send keyEvent with sync.
So we add a new callback to use this API.

Plus,
Widget Application also needs a new API to send an alarm to the WidgetView when it uses a keyEvent.
So we add a new API called SetUsingKeyEvent().

Change-Id: I92c00b57dfe1c72baadb9fc082e253b1a0385d14