[AT-SPI] Refactor Accessibility::BitSet
BitSets is renamed to BitSet (like std::bitset or java.util.BitSet) and
split into two parts: BitSet and EnumBitSet. This way, the enumeration
type can be effectively erased, so that EnumBitSet<SomeEnum, 13> and
EnumBitSet<OtherEnum, 30> can share the same copy of the code in the
resulting binary. EnumBitSet also calculates the required storage size
automatically based on the maximum enum value.

Additionally, the BitSet class is moved from accessibility.h (which
contains mostly enum classes) to a separate file and is enriched with
extensive documentation.

Change-Id: Ib79a370fa9e6241402acb00e859ba20760dbb88c