VectorImageRenderer: Fix data released in case of async loading
If thorvg is used by multiple processes it can't be guaranteed that particular
amount of threads will be used. Especially if we choose to work synchronously,
another process may force it to run asynchronously.
This patch changes copy parameter of load() to true, so to assure that in case
of asynchronous execution the data will not be released before the vector
image loaded.
It also adds printing of the return code in case of mSwCanvas->draw() failure.

Change-Id: I539323768d9fa0be0aeb165b161086973efddd8b