Cache default fontDescription + use default when fontDescription empty
When font description is empty, we will validate font in ValidateFont function.
And inside of that function, fontDescription become mDefaultDescription.

Due to the validateId is not matched, we always fail to get cached font info.
This patch make realFontDescription = mDefaultDescription
when input fontDescription family information is broken, so we can use
font cache feature well.

And also, ValidateFont API runtime is too long. So when we get default
font description, caching it immediatly. It will make that normal case
don't call ValidateFont API.

Change-Id: Iacb8ff84f159da5ee0d80f90c47001da54035e7e
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>