[Tizen] Supports to re-calculate window position and size for window rotation
Supports to re-calculate window position and size when partial size window is rotated.
It is needed when the rotated window is resized or moved.
Because window system APIs work on the default coordinate system
and application sets the window position and size on current orientation.
In addtion, fix the bug invalid touch coordinate when window's size is partial.

Change-Id: I34e861fa34c987db36aa15a9dec26481e80e1f04