[AT-SPI] Implement Component for ProxyAccessible
The Component interface is required for the object not to be immediately
filtered out by bridge-side navigation code. When a ProxyAccessible is
serialized to an object address as a return value on D-Bus, its address
is replaced by the remote address (stored in `mAddress`), so any further
interaction by AT clients (D-Bus method calls) will be directed to the
actual remote object.

Change-Id: I268e219eb30ecd016ab2ab78dcd0846b551a9e42