[AT-SPI] remove isRoot param from Accessible::Get
The Accessibility::Accessible::Get for a window could be called
before getting accessible to add window accessible.

So far, we thought that the next part creates an accessible for
window for the first time.

  accessible = Accessibility::Accessible::Get(rootLayer, true);
  bridge->AddTopLevelWindow(accessible);

However, there is a case where it is created before this part
caused by following symbol

  CSharp_Dali_Toolkit_DevelControl_GetAccessibilityStates+0x144

In this case, isRoot is set to `false`. Then window will have
incorrect accessible information.

For more information, please refer to the following.

  https://github.sec.samsung.net/tizen/atspi/issues/60

Change-Id: Id2d27f35426e72b67986f132f0c77979016b4252