[AT-SPI] Add Bridge::SetPreferredBusName()
This API allows an application to register a string identifier as a bus
name alias on the AT-SPI bus (e.g. "org.tizen.AppName") that remains
constant across a series of activations and deactivations of AT-SPI, in
contrast to the numeric bus name, which changes every time (e.g.
":1.23", ":1.27", ":1.31" etc.).

Change-Id: Idaec55df4bcd73bc26befa5c03c3c1c9e3972781