[AT-SPI] emit showing event for window
For the multi-window case, we need to send 'showing' event when
the AT-SPI bridge turns on. So far, we have sent only 'activate' event
which is enough for the non-multi-window case.

And the 'mIsShown' does not care about the multi-window case.
If one window calls WindowShown, then other windows cannot send 'showing'
event because of the 'mIsShown'.

Change-Id: I501953e8dc1475550afc357293f21dd8fcc98620