[AT-SPI] Add Socket::SetOffset() and related APIs
These APIs allow to set an offset that is added to every GetExtents()
result. The parent process calls Bridge::SetSocketOffset() which makes a
D-Bus call resulting in Bridge::SetExtentsOffset() being called in the
child process (widget). The widget needs this information to report its
screen coordinates correctly to AT-SPI clients, and for navigation to
work.

BridgeComponent::GetExtents adds the value of mExtentsOffset to
Component::GetExtents when reporting back to AT clients. So, in the
reverse operation of GetNavigableAtPoint, the value of mExtentsOffset
has to be subtracted.

Change-Id: I4e563c8be1267101c355e27f6e2782f62655347f