Cache rendered glyph if required. + Copyless glyph creation
Call FT_Glyph_To_Bitmap might be overhead for some system.
But cache the rendered bitmap information could be another overhead.

So, we try to ask environment and store the bitmap result.

Follow as EFL system, we will store it as 8bit -> 4bit reduced
and RLE4 compression.

+

Now we can know the buffer's ownership for FontClient::GlyphBufferData.
We can use the buffer without copy if need.

Change-Id: I82a76961c84679c04750f0de904096130cb69df2
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>