Add UI thread feature on Tizen
The UiThread means that the application creates a thread for the UI events.
So the Dali::Adaptor is created on the UiThread, not the main thread.
When you enable UiThread,
the signals (Init, Terminate, Pause, Resume, Reset, AppControl, LanguageChanged, RegionChanged, LowBattery, LowMemory)
are emitted on the UiThread.
If you want to get the events on the main thread, you must connect the task signals(TaskInit, TaskTerminate, etc.).
When you don't enable the UiThread, the task signals aren't emitted.

Change-Id: I2cc57e85fa50f9b854531efa6048796aadbe6da9