[AT-SPI] Suppress some events in Widget windows
While the Widget is technically a window in DALi, it should not be seen
as a window by AT clients such as the Screen Reader, i.e. it should not
grab and narrow down the navigation context. Suppressing window events
and some state events (notably, State::SHOWING, however the API offers
an all-or-nothing granularity only) is necessary to achieve this goal.

Change-Id: I0873751f025bbdd321aa26444c070bf064c0ce3f