[AT-SPI] De-duplicate serialization code for Accessible
(De)serializing an Accessible is almost entirely equivalent to
(de)serializing an Address, so most code can be reused. This patch also
fixes a crash that would occur in libdbus when Accessible* was non-null,
but its address was empty (could be the case for ProxyAccessible),
yielding an invalid object path that would trigger an assertion.

Change-Id: I595c42d5257edc2052d18927f5e08650efef8417