[AT-SPI] Do not store the window in mDefaultLabels
The default label stack was often incorrect when Screen Reader was
turned on dynamically, due to how window registration was handled. There
is no need to store the window on the default label stack explicitly at
all, because the navigation context root is also the target of the
incoming D-Bus call. The broadest possible context is the window, which
is the case if the navigation context stack is empty.

Change-Id: I61d4f279c1c0f82c75516359db73cd941f5d0b53