(Capture) Don't scene-off camera when we use inputed camera
If we create capture with our own camera, App can add/remove this
camera actor.

But current capture API automatically dettach the camera.

If we use DALi's default camera, than this default camera dettached
from the window, and the rendering broken.

This patch protect some kind of this scenario.
We will dettach the camera from window only if we use
capture-impl generated camera, and don't dettach if we use
inputed camera.

Change-Id: I902fbe3a4d933165de87cff7d420438e66c2fdd2
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>