[AT-SPI] Prevent flooding D-Bus with PostRender events
When enabled with SetListenPostRender, the window Accessible emits countless
PostRender events on D-Bus. The CoalescableMessages mechanism can be used here
just like for BoundsChanged and SetOffset. This way, with this patch, the
PostRender events come at a steady rate. With CoalescableMessages at least one
event will still be emitted if requested, which is important for clients who
may depend on there being a PostRender event on the AT-SPI bus after a
particular operation -- there still is, but without the flooding.

Change-Id: Iab9451d3bff7bf57007388ed641ba543d40b52c1