Fix performance down if DefaultFontDescription not prepared
Since FontDescription cache system have some problem if family is empty.
So we use DefaultFontDescription if inputed family is empty.

But we try to delay the FontClient creation time as possible.
If that case, It it possible that Default informations are not prepared yet.

This patch ensure the mDefaultFontDescription prepared at FontClient::Get().
So now, we can ensure to use FontId cache system well.

TODO : We might fix empty family bugs if possible

Change-Id: I72f5f7a581186addd130848f257f939cd24e094a
Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>