Add font thread sync creation option
If sync creation is true, font thread creation guarantees syncronization with the main thread.

The main thread runs the font thread and waits,
and it wakes up upon receiving a notification from the font thread.
If it doesn't receive a notification within the specified time, it wakes up due to a timeout.

test security manager sync thread issue

Change-Id: I59ef4c189f0765cfd1c955704744c219784d238b
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>