[Tizen][AT-SPI] Allow manual control of ReadingMaterial::listChildrenCount
In complex UIs, where the logical number of children (i.e. not the number of
Actors) is difficult to determine, because (1) the children have differing
roles, or (2) the children are not direct descendants of the parent object, the
best-effort calculation done by the AT-SPI bridge (for the purpose of reading
"Showing %d items") will often be wrong.

This patch recognizes the "item_count" attribute as a high-priority, manual
setting of the number of children for the purpose of reading "Showing %d
items" with the provided value.

Change-Id: I2d027ea4930d7e9f63dee9fc4bb55a14caebf7b8