Fixed network client to trigger main thread
Main thread Trigger should be created on the event
thread, not in client threads.

Caused problems with the trigger implementation, as that stored
main loop context in thread local storage, so was always 0 / NULL,
and this caused the network client to crash or do nothing,
depending on the chosen main loop implementation.

Change-Id: I483cb1dc55b1ff4f6cf81434780bcd560f2f57a9
Signed-off-by: David Steele <david.steele@samsung.com>