Modify function description in window.
<Summary>
Windows modules are not handled by one developer, but by multiple developers in various fields. As a result, coding rules and function descriptions are poorly written compared to other modules.
This patch is a supplement to these parts.

<What>
The function description of the header file of the window-impl and window render surface, which are common modules of the window, has been checked and modified.
the format is as follows.
1. Internal function
{
@brief

@param[in]
@return
}
2. override function
{
@copydoc
}

1 internal function is described, and the contents  description of function description, @param[in] input parameter,  description of return value.
Number 2 describes the overridden function, and the main description is defined in the header file of the inherited class.

<Why>
As the DALi of the Rendering Engine is changed from the Public Interface to the internal module, the description creation or management in the header file is becoming poor.
This is to correct this, increase readability, and strengthen communication.

<Plan>
Currently, we have only performed for window common modules, but we plan to communicate with the person in charge so that they can be applied to other modules in Dali in the future.

Change-Id: Ifa04220d176dde50cdc6132826019a99f1021491