The template classes should not have the tag of module.
the APIs of tempalte class will be exapned in the module who use it,
so it should not have the tag of module.

For example:

The class "BitSet" is used in dali-toolkit, then the APIs of the class are defined in dali-toolkit,
so it is not a class of dali-adaptor.

Change-Id: Ieb53a42f90d39bf34f6a75b4c2d0e13d4e7e19ab