Ensure to check whether task need to callback or not
There was short-term to unlock the mRunningTasksMutex.
If main thread Remove the running task (mean, cancel) during
that mutex unlocked, we just keep know that
this task need to execute callback.

So, we need to check for notify or not only for lottie cases,
and check main thread callback required one more time for main thread cases.

Change-Id: I8e6bdcdf4b8bf290a23b5b8e3fdb1821db26d393
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>