[Tizen] Fix ellipsis cache issue
If an ellipsis cache is not found in the mEllipsisCache,
the ellipsisCacheIndex value will always be zero.

This zero value is then stored in the item.index,
which is used to access the mEllipsisCache.

As a result, even if an ellipsis cache exists,
only the cache at index 0 is always used,
so an ellipsis glyph with a different font size from the actual font size may be rendered.

Change-Id: Iee15f1e320fd3f8ba73e035477ed13d7e49ee0ef
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>