[Tizen] Fixing static analysis error
The libturbojpeg library has been updated on the target, and there is
a new subsampling option available, which means the array we use for
testing is now too small.

Added an entry to this array for the new subsampling option. In
theory, this should just work, but as we don't set this variable,
there's nothing to test.

See
https://review.tizen.org/gerrit/c/platform/core/uifw/dali-adaptor/+/277590
for more details

Change-Id: I02bb90830d6d0969cafae635ceb6b1b6bc9c2e1a
Signed-off-by: David Steele <david.steele@samsung.com>