Changed FBOs to use EglSync
For texture dependencies between shared context and render surface
context (i.e. Window/Scene surface), e.g. when an offscreen is
rendered on-screen, have changed to use EglClientWaitSync with
timeout = FOREVER.

This blocks the CPU when sending draw calls to the GPU where
that draw call uses an offscreen texture. However, in testing
gaussian-blur-view.example, this lasts approx 10ms on some frames
and 16/17 ms on others, yielding an FPS of 41.

Changed egl-sync-implementation.cpp to have multiple versions
of the file for each platform, rather than old-style #ifdefs.

(Todo: could check if the offscreen was re-rendered this frame, as
there is no need to sync if it hasn't changed. But, client can use
RENDER_ONCE to force this behaviour).

Change-Id: I64eaf82c5203e6acb6e0fc76b0f0e36136defcdf