[TIZEN] Check if child is visible in scrollable parent.
Fixes an issue where GetNeighbor function would incorrectly find the next item in a gridview, even if child item was not actually visible.
In a grid view, the next item may be located to the right of the current item, even if it is not visible on the screen due to scrolling.
Add IsChildVisibleInScroll to checking the next item is actually visible or not.
This change improves the user experience by ensuring that the highlight function always finds the correct next item, even in grid views.

Change-Id: Id9cbe3f97d29e7a8d5458a6b60e6d7f9ab0f27a6