Changed render target surface to RenderSurface i/f
The surface is alive before the graphics surface is created on it,
and we need a valid surface to send to core. Modern GLES
implementation uses a void* and passes the RenderSurfaceInterface
through this pointer.

This version of DALI still has Integration::RenderSurface in
dali-core, so have changed to use this.

Remove the Graphics::Surface, this doesn't map to anything useful in
dali-core. Kept the Graphics::FramebufferId, but this is only useful
in dali-adaptor side (matching window to graphics surface/swapchain
pair).

Signed-off-by: David Steele <david.steele@samsung.com>