Changed test harness to work with Vulkan build
Renamed dali-graphics module to dali-egl-graphics

If the main build is Vulkan, then some of the test modules
won't build, nor will some of the tests in particular modules.

Added a configure option to the main build to configure the
top level test harness CMakeLists.txt with the ENABLE_VULKAN option.

Added handling to the scripts to use CMake cache to figure
out list of source files for a given module rather than using
AWK on CMakeLists.txt. This enables conditional source inclusion
based on whether Vulkan is enabled or not in the main build.
(tct-<module>-core.h is generated in source tree by these scripts)

Changed the build script to ignore dali-egl-graphics module
if VULKAN is enabled.

Changed the execute script to use only the built folders, not the
source folders to determine what to run.

Change-Id: Id00dd7c00d011af7d40c6b4b1b138e0ac0a58fa6
Signed-off-by: David Steele <david.steele@samsung.com>