Remove list of egl context what egl hold
Since there was no way to remove mEglWindowContext, let we remove it
when render-surface-interface call it.

Note : The only user who don't call DestroyContext is native-render-surface-ecore-wl.cpp,
and drawable-view-native-renderer.cpp

We need to check the best way to detroy context for it now.
But here, we should remove them, for the general cases

Change-Id: I0329d50b2a75ea7b9c645673feb81e4aeca6523a
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>