Fix asterisk count mismatch in password mode
Until now, the number of asterisks '*' has been used based on the text length..
instead of the number of utf8 characters in password mode.

This caused an issue where the number of rendered asterisks did not match the number of asterisks delivered to the IME.
Now, when creating an asterisk string to deliver to the IME,
we use the number of utf8 characters.

Change-Id: I76324f2ed9326102196441d72c8a96d59fce792d
Signed-off-by: Bowon Ryu <bowon.ryu@samsung.com>