Ensure render once after we set surface full swap
It is possible if event thread running too slow, the surface's full swap infomatin
might not be applied to update render thread.

For example : When OnIconifyChanged come, mAdaptor->Resume() called so
combined update render thread wake up.
But it is possible that window render surface's SetFullSwapNextFrame() not be called
until the end of render loop.
If then, the full scene rendering might be skiped after Adaptor resume finished.

To avoid this kind of issue, let we request update once when
SetFullSwapNextFrame() called after ProcessCoreEvents().

Change-Id: I40e738f6a7d56f1abd3e6d0dec37ad0de5a57994
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>